Added framework for test suite, beginning with utilities.py
Starting to arrange a test suite for the SAR project; long term
goal of which is to have a system in place that should guarantee
that no non-functional code can be committed to the master branch.

This begins using the Python unittest module, but I might change
to a different one as soon as I get my actual environment up and
running. For much the same reason, this is untested and is being
kept in my personal branch.